Feature: useBridge hook
